feat: add `SignatureDownlaoder.kt` * Added automatic signatures downloader * Refactor `utils` package.
